requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
bert-score==0.3.13
bitsandbytes==0.45.3
blis==0.7.11
Brotli @ file:///croot/brotli-split_1714483155106/work
catalogue==2.0.10
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
datasets==3.1.0
deepspeed==0.13.4
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
hjson==3.1.0
huggingface-hub==0.29.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.36.0
iniconfig==2.1.0
Jinja2 @ file:///croot/jinja2_1730902924303/work
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.8.4
mdurl==0.1.2
medspacy==1.3.1
medspacy_quickumls==3.2
medspacy_unqlite==0.9.6
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
monai==1.3.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
networkx @ file:///croot/networkx_1720002482208/work
nibabel==5.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
oauthlib==3.2.2
open-clip-torch==2.24.0
opencv-python==4.9.0.80
openpyxl==3.1.5
packaging==24.2
pandas==2.2.2
peft==0.14.0
pillow @ file:///croot/pillow_1731594689713/work
platformdirs==4.3.7
pluggy==1.5.0
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycocotools==2.0.7
pydantic==2.9.2
pydantic_core==2.23.4
pydicom==3.0.1
PyFastNER==1.0.10
Pygments==2.19.1
pynvml==11.5.3
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
PyRuSH==1.0.9
pysbd==0.3.4
pysimstring==1.3.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
qimage2ndarray==1.10.0
QtAwesome==1.3.1
QtPy==2.4.2
quicksectx==0.4.0
RaTEScore==0.5.0
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib==1.3.1
responses==0.18.0
rich==14.0.0
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.24.0
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.13.0
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
shellingham==1.5.4
simple-slice-viewer==0.97
SimpleITK==2.3.1
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy @ file:///croot/sympy_1724938189289/work
tensorboardX==2.6.2.2
thinc==8.2.5
threadpoolctl==3.6.0
timm==1.0.11
tokenizers==0.21.0
tomli==2.2.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.67.0
transformers==4.49.0
triton==2.2.0
tweepy==4.14.0
typer==0.15.3
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
Unidecode==1.4.0
urllib3 @ file:///croot/urllib3_1727769808118/work
wandb==0.19.9
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.17.1
